idk whats diffrent nir you go check

+++ functions.py
@@ -24,7 +24,12 @@ def dir_cmd(msg):
     """
     output = ERR
     try:
-        output = ','.join(glob.glob(msg[1] + '\\*.*'))
+        if(len(msg) < 2):
+             output = 'pls enter the all reqiured parm'
+        else: 
+            output = ','.join(glob.glob(msg[1] + '\\*.*'))
+            if(output == ''):
+                output = 'directory doesnt exist'
     except Exception as err:
         output += ' ' + str(err)
     finally:
@@ -42,8 +47,11 @@ def delete_cmd(msg):
     """
     output = ERR
     try:
-        os.remove(msg[1])
-        output = 'file deleted successfully'
+        if(len(msg) < 2):
+             output = 'pls enter the all reqiured parm'
+        else: 
+            os.remove(msg[1])
+            output = 'file deleted successfully'
     except Exception as err:
         output += ' ' + str(err)
     finally:
@@ -61,8 +69,11 @@ def copy_cmd(msg):
     """
     output = ERR
     try:
-        shutil.copy(msg[1], msg[2])
-        output = 'file copied successfully'
+        if(len(msg) < 3):
+             output = 'pls enter the all reqiured parm'
+        else: 
+            shutil.copy(msg[1], msg[2])
+            output = 'file copied successfully'
     except Exception as err:
         output += ' ' + str(err)
     finally:
@@ -80,8 +91,11 @@ def execute_cmd(msg):
     """
     output = ERR
     try:
-        subprocess.call(msg[1])
-        output = 'command executed successfully'
+        if(len(msg) < 2):
+             output = 'pls enter the all reqiured parm'
+        else: 
+            subprocess.call(msg[1])
+            output = 'command executed successfully'
     except Exception as err:
         output += ' ' + str(err)
     finally:
@@ -108,9 +122,19 @@ def take_screenshot_cmd(msg):
         return output
 
 def send_photo_cmd(msg):
+    """
+    Sends screenshot.
+
+    :param msg: The command and the directory path.
+    :type msg: list
+
+    :return: image as base64 encode as astring.
+    :rtype: str
+    """
     output = ERR
     try:
         with open('screen.jpg', "rb") as image_file:
+            #encode image to base64 as astring
             encoded_string = base64.b64encode(image_file.read()).decode('utf-8')
         output = encoded_string
     except Exception as err:
